♻️ refactor(test): address PR review comments
- Extract shared SettingsScope to TestHelpers/SettingsScope.cs
  (was duplicated verbatim in AppearanceViewModelTests,
  StorageViewModelTests and BackupViewModelTests)
- Remove redundant out ICategoryRepository parameter from
  CategoriesViewModelTests.CreateSut — callers already hold the
  reference; DeleteKategorie_WhenNotConfirmed_DoesNotDelete now
  declares a local variable instead

All 195 tests still pass.
